// frontend/src/lib/types/enums/user-permissions-enum.ts
export enum UserPermissionsEnum {
  AUTH_TOTP_READ = "AUTH_TOTP_READ",
  AUTH_TOTP_CREATE = "AUTH_TOTP_CREATE",
  AUTH_TOTP_DELETE = "AUTH_TOTP_DELETE",

  APP_SETTINGS_READ = "APP_SETTINGS_READ",
  APP_SETTINGS_UPDATE = "APP_SETTINGS_UPDATE",

  USERS_READ = "USERS_READ",
  USERS_CREATE = "USERS_CREATE",
  USERS_UPDATE = "USERS_UPDATE",
  USERS_DELETE = "USERS_DELETE",

  SENDERS_READ = "SENDERS_READ",
  SENDERS_CREATE = "SENDERS_CREATE",
  SENDERS_UPDATE = "SENDERS_UPDATE",
  SENDERS_DELETE = "SENDERS_DELETE",

  CLIENTS_READ = "CLIENTS_READ",
  CLIENTS_CREATE = "CLIENTS_CREATE",
  CLIENTS_UPDATE = "CLIENTS_UPDATE",
  CLIENTS_DELETE = "CLIENTS_DELETE",

  CLIENT_TOKENS_READ = "CLIENT_TOKENS_READ",
  CLIENT_TOKENS_CREATE = "CLIENT_TOKENS_CREATE",
  CLIENT_TOKENS_DELETE = "CLIENT_TOKENS_DELETE",

  TEMPLATES_READ = "TEMPLATES_READ",
  TEMPLATES_CREATE = "TEMPLATES_CREATE",
  TEMPLATES_UPDATE = "TEMPLATES_UPDATE",
  TEMPLATES_DELETE = "TEMPLATES_DELETE",

  EMAILS_READ = "EMAILS_READ",
  EMAILS_CREATE = "EMAILS_CREATE"
}